Routes are now reused and less need of strings.
So now the index view can redirect to a page using the actual page
class, not a string. So we can change the route in one place and it will
fall through to the rest.

Also made it so pages should cache their route object, should be more
efficient and prevent double routing.